papermage_components/chem_data_extractor_predictor/requirements.txt
chemdataextractor2
fastapi~=0.103.2
uvicorn[standard]